new setup of lsp-zero and mason
