+++ nvim/lua/colorscheme.lua
@@ -1,0 +1,4 @@
+-- Colorshceme
+vim.o.background = "dark" -- or "light" for light mode
+vim.cmd("colorscheme gruvbox")
+vim.cmd("hi MatchParen cterm=bold ctermbg=none ctermfg=red")
