+++ nvim/lua/settings.lua
@@ -1,0 +1,23 @@
+-- general vim Settings
+
+vim.opt.encoding="utf-8"
+vim.opt.number=true
+
+--- Indentation
+vim.opt.shiftwidth=4
+vim.opt.softtabstop=4
+vim.opt.tabstop=4
+vim.opt.expandtab=true
+vim.opt.smartindent=true
+
+--- Splitting
+vim.opt.splitbelow=true
+vim.opt.splitright=true
+
+vim.opt.wrapscan=false
+vim.opt.ignorecase=true
+vim.opt.textwidth=80
+vim.opt.signcolumn="yes"
+
+-- Scroll earlier
+vim.o.scrolloff = 8;
